checkpoint - migrate to using long/Long to represent any integer ID value (rowId, objectId, etc).

+++ lincs/src/org/labkey/lincs/DocImportListener.java
@@ -137,7 +137,7 @@ public class DocImportListener implements ExperimentListener, SkylineDocumentImp
             return;
         }
 
-        int jobId = PipelineService.get().getJobId(user, container, job.getJobGUID());
+        long jobId = PipelineService.get().getJobId(user, container, job.getJobGUID());
         _log.info("LINCS: Queued job Id " + jobId +" for creating GCT files for " + skylineRun.getFileName() + ". Container: " + container.getPath());
 
         pspJob.setPipelineJobId(jobId);
